+++ .github/jobs/sanitize.sh
@@ -7,10 +7,18 @@ set -xeuo pipefail
 sanitize_runs=(
   "ctor:ctor-example"
   "ctor:ctor-advanced"
+  "link-section:link-section-dyn"
   "link-section:link-section-example"
-  "link-section:link-section-ref"
   "link-section:link-section-mut"
+  "link-section:link-section-mut-no-macro"
   "link-section:link-section-movable"
+  "link-section:link-section-movable-no-macro"
+  "link-section:link-section-ref"
+  "scattered-collect:scattered-collect-intern-strings"
+  "scattered-collect:scattered-collect-slice"
+  "scattered-collect:scattered-collect-sorted-slice"
+  "scattered-collect:scattered-collect-referenced-slice"
+  "scattered-collect:scattered-collect-sorted-referenced-slice"
 )
 
 for spec in "${sanitize_runs[@]}"; do